# lexical-search/openapi/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Monty Lexical Search API",
    "description": "A REST API for building and querying lexical search indices",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://0.0.0.0:42059/api/v1"
    }
  ],
  "security": [],
  "paths": {
    "/workspaces": {
      "get": {
        "summary": "Returns a list of workspaces",
        "operationId": "listWorkspaces",
        "parameters": [
          {
            "name": "workTree",
            "in": "query",
            "description": "The workspace path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new workspace",
        "operationId": "createWorkspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a workspace if it exists",
        "operationId": "deleteWorkspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/updates": {
      "get": {
        "summary": "Stream archive updates (SSE) for a workspace",
        "operationId": "streamUpdates",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The workspace id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "A list of Monty databases",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MontyDatabase"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Workspace": {
        "type": "object",
        "required": [
          "id",
          "name",
          "bleveDir",
          "gitDir",
          "workTree",
          "docCount",
          "archiveDBVersion",
          "lastIndexedCommit",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The workspace id"
          },
          "name": {
            "type": "string",
            "description": "The workspace name"
          },
          "bleveDir": {
            "type": "string",
            "description": "The abs path of the bleve index"
          },
          "gitDir": {
            "type": "string",
            "description": "The abs path of the .git folder"
          },
          "workTree": {
            "type": "string",
            "description": "The abs path of the workspace folder"
          },
          "docCount": {
            "type": "integer",
            "description": "The number of documents in the bleve index"
          },
          "archiveDBVersion": {
            "type": "integer",
            "description": "The version of the archive database"
          },
          "lastIndexedCommit": {
            "type": "string",
            "description": "The commit hash of the last successfully indexed commit"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WorkspacesResponse": {
        "type": "object",
        "required": [
          "workspaces"
        ],
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "links": {
            "type": "object",
            "properties": {
              "about": {
                "type": "string"
              }
            }
          },
          "status": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "type": "string"
          }
        }
      },
      "ErrorsResponse": {
        "type": "object",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name",
          "workTree"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The workspace name"
          },
          "workTree": {
            "type": "string",
            "description": "The absolute workspace path",
            "example": "/path/to/workspace"
          }
        }
      },
      "DeleteWorkspaceRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The workspace id"
          }
        }
      },
      "MontyDatabaseUsageMode": {
        "type": "string",
        "enum": [
          "always",
          "onRequest"
        ]
      },
      "MontyDatabase": {
        "type": "object",
        "required": [
          "name",
          "folders",
          "usageMode"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the Monty database"
          },
          "folders": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A list of folders associated with the Monty database"
          },
          "usageMode": {
            "$ref": "#/components/schemas/MontyDatabaseUsageMode",
            "description": "The usage mode of the Monty database"
          }
        }
      },
...
    }
  }
}